requirements.txt
pandas~=2.2.2
numpy~=1.26.4
matplotlib~=3.9.0
beautifulsoup4~=4.12.3
requests~=2.32.3
plotly~=5.14
chess~=1.10.0

pillow~=10.4.0
scipy~=1.13.1